Permissions added

+++ events/messageCreate.js
@@ -2,24 +2,45 @@ const { PrismaClient } = require('@prisma/client');
 const prisma = new PrismaClient();
 const  { MessageEmbed } = require('discord.js');
 
+const escapeRegex = (str) => str.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
+
 module.exports = async (client, message) => {
     if (!message.guild ||
         message.author.bot) return;
 
+        const prefix = "!"
+
+        const prefixRegex = new RegExp(`^(${escapeRegex(prefix)})\\s*`);
+
+        if (!prefixRegex.test(message.content)) return;
+
+        const [ matchedPrefix ] = message.content.match(prefixRegex);
+
+        const args = message.content.slice(matchedPrefix.length).trim().split(/ +/g); 
+
         const textExists = await prisma.cCommands.findUnique({
             where: {
-                id: message.content
+                placeholder: `${args[0] + message.guild.id}`
             },
             select: {
                 text: true,
                 guild: true,
-                deltrig: true
+                deltrig: true,
+                permission: true
             }
         })
 
         if(textExists) {
 
             if(textExists.guild === message.guild.id) {
+
+                if(textExists.permission === "staff") {
+                    if(!message.member.permissions.has("MANAGE_MESSAGES")) return;
+                }
+
+                if(textExists.permission === "admin") {
+                    if(!message.member.permissions.has("ADMINISTRATOR")) return;
+                }
 
                 if(textExists.deltrig === true) {
                     message.delete();
@@ -35,17 +56,26 @@ module.exports = async (client, message) => {
 
         const embedExists = await prisma.cCEmbeds.findUnique({
             where: {
-                id: message.content
+                placeholder: `${args[0] + message.guild.id}`
             },
             select: {
                 description: true,
                 guild: true,
                 title: true,
-                deltrig: true
+                deltrig: true,
+                permission: true
             }
         })
 
         if(embedExists) {
+
+            if(embedExists.permission === "staff") {
+                if(!message.member.permissions.has("MANAGE_MESSAGES")) return;
+            }
+
+            if(embedExists.permission === "admin") {
+                if(!message.member.permissions.has("ADMINISTRATOR")) return;
+            }
 
             if(embedExists.guild === message.guild.id) {
                 let embed = new MessageEmbed()
